Add support for half-precision types in Metal.
This will hopefully improve performance on lower-end GPUs.

Bug: skia:12339
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/465078
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>

+++ tests/sksl/intrinsics/BitCount.metal
@@ -8,12 +8,12 @@ struct Uniforms {
 struct Inputs {
 };
 struct Outputs {
-    float4 sk_FragColor [[color(0)]];
+    half4 sk_FragColor [[color(0)]];
 };
 fragment Outputs fragmentMain(Inputs _in [[stage_in]], constant Uniforms& _uniforms [[buffer(0)]], bool _frontFacing [[front_facing]], float4 _fragCoord [[position]]) {
     Outputs _out;
     (void)_out;
-    _out.sk_FragColor.x = float(popcount(_uniforms.a));
-    _out.sk_FragColor.y = float(popcount(_uniforms.b));
+    _out.sk_FragColor.x = half(popcount(_uniforms.a));
+    _out.sk_FragColor.y = half(popcount(_uniforms.b));
     return _out;
 }
