+++ tests/sksl/intrinsics/Fma.metal
@@ -2,14 +2,14 @@
 #include <simd/simd.h>
 using namespace metal;
 struct Uniforms {
-    float a;
-    float b;
-    float c;
+    half a;
+    half b;
+    half c;
 };
 struct Inputs {
 };
 struct Outputs {
-    float4 sk_FragColor [[color(0)]];
+    half4 sk_FragColor [[color(0)]];
 };
 fragment Outputs fragmentMain(Inputs _in [[stage_in]], constant Uniforms& _uniforms [[buffer(0)]], bool _frontFacing [[front_facing]], float4 _fragCoord [[position]]) {
     Outputs _out;
