+++ tests/sksl/metal/CastMat4x4ToMat3x4.metal
@@ -4,7 +4,7 @@ using namespace metal;
 struct Inputs {
 };
 struct Outputs {
-    float4 sk_FragColor [[color(0)]];
+    half4 sk_FragColor [[color(0)]];
 };
 float3x4 float3x4_from_float4x4(float4x4 x0) {
     return float3x4(float4(x0[0].xyzw), float4(x0[1].xyzw), float4(x0[2].xyzw));
@@ -14,6 +14,6 @@ fragment Outputs fragmentMain(Inputs _in [[stage_in]], bool _frontFacing [[front
     (void)_out;
     float3x4 a = float3x4(1.0);
     float3x4 b = float3x4_from_float4x4(float4x4(1.0));
-    _out.sk_FragColor.x = float(all(a[0] == b[0]) ? 0 : 1);
+    _out.sk_FragColor.x = half(all(a[0] == b[0]) ? 0 : 1);
     return _out;
 }
