+++ tests/sksl/metal/CastMat4x4ToMat4x3.metal
@@ -4,7 +4,7 @@ using namespace metal;
 struct Inputs {
 };
 struct Outputs {
-    float4 sk_FragColor [[color(0)]];
+    half4 sk_FragColor [[color(0)]];
 };
 float4x3 float4x3_from_float4x4(float4x4 x0) {
     return float4x3(float3(x0[0].xyz), float3(x0[1].xyz), float3(x0[2].xyz), float3(x0[3].xyz));
@@ -14,6 +14,6 @@ fragment Outputs fragmentMain(Inputs _in [[stage_in]], bool _frontFacing [[front
     (void)_out;
     float4x3 a = float4x3(1.0);
     float4x3 b = float4x3_from_float4x4(float4x4(1.0));
-    _out.sk_FragColor.x = float(all(a[0] == b[0]) ? 0 : 1);
+    _out.sk_FragColor.x = half(all(a[0] == b[0]) ? 0 : 1);
     return _out;
 }
