+++ tests/sksl/shared/Offset.metal
@@ -9,13 +9,13 @@ struct Test {
 struct Inputs {
 };
 struct Outputs {
-    float4 sk_FragColor [[color(0)]];
+    half4 sk_FragColor [[color(0)]];
 };
 fragment Outputs fragmentMain(Inputs _in [[stage_in]], bool _frontFacing [[front_facing]], float4 _fragCoord [[position]]) {
     Outputs _out;
     (void)_out;
     Test t;
     t.x = 0;
-    _out.sk_FragColor.x = float(t.x);
+    _out.sk_FragColor.x = half(t.x);
     return _out;
 }
