+++ tests/sksl/shared/StaticSwitchWithFallthroughA.metal
@@ -4,7 +4,7 @@ using namespace metal;
 struct Inputs {
 };
 struct Outputs {
-    float4 sk_FragColor [[color(0)]];
+    half4 sk_FragColor [[color(0)]];
 };
 fragment Outputs fragmentMain(Inputs _in [[stage_in]], bool _frontFacing [[front_facing]], float4 _fragCoord [[position]]) {
     Outputs _out;
@@ -14,6 +14,6 @@ fragment Outputs fragmentMain(Inputs _in [[stage_in]], bool _frontFacing [[front
         x = 0.0;
         x = 1.0;
     }
-    _out.sk_FragColor = float4(x);
+    _out.sk_FragColor = half4(half(x));
     return _out;
 }
